+++ R/read_tern.R
@@ -47,7 +47,9 @@
 #'     `"SMindex"`, `"bucket1"`, `"bucket2"`,
 #'     `"deepD"`, or `"runoff"`.}
 #' }
-#' Data availability: from 2005-01-01 to today. (Updated regularly.)
+#' Data availability: `totalbucket` and `SMindex` from 2005-01-01;
+#' `bucket1`, `bucket2`, `deepD`, and `runoff` from 2015-01-01; to today.
+#' (Updated regularly.)
 #'
 #' @section ASC -- Australian Soil Classification (`"ASC"`):
 #' \describe{